// VocaDbWeb/New/types/Models/Artists/ArtistCategories.ts
// Corresponds to the ArtistCategories enum in C#.
export enum ArtistCategories {
	Nothing = 'Nothing',
	Vocalist = 'Vocalist',
	Producer = 'Producer',
	Animator = 'Animator',
	Label = 'Label',
	Circle = 'Circle',
	Other = 'Other',
	Band = 'Band',
	Illustrator = 'Illustrator',
	Subject = 'Subject',
}